+++ packages/abstract-extractor/lib/ergo/initializers/ergoBoxInitializer.ts
@@ -1,7 +1,7 @@
 import { Mutex } from 'await-semaphore';
 import { groupBy, sortBy } from 'lodash-es';
 
-import { DummyLogger } from '@rosen-bridge/abstract-logger';
+import { DummyLogger, AbstractLogger } from '@rosen-bridge/abstract-logger';
 import {
   BlockInfo,
   ErgoNetworkType,
@@ -37,7 +37,7 @@ export class ErgoBoxInitializer<
     ) => Promise<boolean>,
     protected actions: AbstractErgoBoxAction<ExtractedData, ExtractorEntity>,
     maxParallelRequests = MAX_PARALLEL_REQUESTS,
-    protected logger = new DummyLogger(),
+    protected logger: AbstractLogger = new DummyLogger(),
   ) {
     super(
       networkType,