+++ packages/abstract-scanner/lib/scanner/abstract/scanner.ts
@@ -3,14 +3,15 @@ import { difference, remove } from 'lodash-es';
 
 import { AbstractExtractor } from '@rosen-bridge/abstract-extractor';
 import { AbstractLogger, DummyLogger } from '@rosen-bridge/abstract-logger';
+import { ObjectLiteral } from '@rosen-bridge/extended-typeorm';
 import { Block, BlockInfo } from '@rosen-bridge/scanner-interfaces';
 
 import { BlockDbAction } from '../action';
 
 export abstract class AbstractScanner<TransactionType> {
   action: BlockDbAction;
-  extractors: Array<AbstractExtractor<TransactionType>>;
-  newExtractors: Array<AbstractExtractor<TransactionType>>;
+  extractors: Array<AbstractExtractor<TransactionType, ObjectLiteral>>;
+  newExtractors: Array<AbstractExtractor<TransactionType, ObjectLiteral>>;
   logger: AbstractLogger;
   initializeMutex: Mutex;
 
@@ -85,10 +86,10 @@ export abstract class AbstractScanner<TransactionType> {
    * @param extractor
    */
   registerExtractor = async (
-    extractor: AbstractExtractor<TransactionType>,
+    extractor: AbstractExtractor<TransactionType, ObjectLiteral>,
   ): Promise<void> => {
     const notRegisteredIn = (
-      extractors: Array<AbstractExtractor<TransactionType>>,
+      extractors: Array<AbstractExtractor<TransactionType, ObjectLiteral>>,
     ) =>
       extractors.filter(
         (extractorItem) => extractorItem.getId() === extractor.getId(),
@@ -112,9 +113,9 @@ export abstract class AbstractScanner<TransactionType> {
    * @param extractor
    */
   removeExtractor = async (
-    extractor: AbstractExtractor<TransactionType>,
+    extractor: AbstractExtractor<TransactionType, ObjectLiteral>,
   ): Promise<void> => {
-    const removeFn = (ex: AbstractExtractor<TransactionType>) =>
+    const removeFn = (ex: AbstractExtractor<TransactionType, ObjectLiteral>) =>
       ex.getId() === extractor.getId();
 
     const release = await this.initializeMutex.acquire();
@@ -154,7 +155,9 @@ export abstract class AbstractScanner<TransactionType> {
    * @param block
    */
   protected verifyExtractorsInitialization = async (block: BlockInfo) => {
-    const getIds = (extractors: Array<AbstractExtractor<TransactionType>>) => {
+    const getIds = (
+      extractors: Array<AbstractExtractor<TransactionType, ObjectLiteral>>,
+    ) => {
       return extractors.map((extractor) => extractor.getId());
     };
     this.logger.debug(`Initializing extractors for block [${block.height}]`);