+++ packages/watcher-data-extractor/lib/extractor/permitExtractor.ts
@@ -5,16 +5,17 @@ import { difference } from 'lodash-es';
 
 import { AbstractExtractor } from '@rosen-bridge/abstract-extractor';
 import { AbstractLogger, DummyLogger } from '@rosen-bridge/abstract-logger';
-import { DataSource } from '@rosen-bridge/extended-typeorm';
+import { DataSource, SelectQueryBuilder } from '@rosen-bridge/extended-typeorm';
 import { BlockInfo, Transaction } from '@rosen-bridge/scanner-interfaces';
 import ergoExplorerClientFactory, { V1 } from '@rosen-clients/ergo-explorer';
 
 import PermitAction from '../actions/permitAction';
 import { DefaultApiLimit } from '../constants';
+import PermitEntity from '../entities/permitEntity';
 import { ExtractedPermit } from '../interfaces/extractedPermit';
 import { JsonBI } from '../utils';
 
-class PermitExtractor extends AbstractExtractor<Transaction> {
+class PermitExtractor extends AbstractExtractor<Transaction, PermitEntity> {
   readonly logger: AbstractLogger;
   id: string;
   private readonly dataSource: DataSource;
@@ -39,7 +40,10 @@ class PermitExtractor extends AbstractExtractor<Transaction> {
       .to_base16_bytes();
     this.RWT = RWT;
     this.logger = logger ? logger : new DummyLogger();
-    this.actions = new PermitAction(dataSource, this.logger);
+    this.actions = new PermitAction(
+      dataSource,
+      this.logger.child('PermitAction'),
+    );
     this.explorerApi = ergoExplorerClientFactory(explorerUrl);
   }
 
@@ -285,6 +289,15 @@ class PermitExtractor extends AbstractExtractor<Transaction> {
     }
     return extractedBoxes;
   };
+
+  /**
+   * Builds a query that returns used blocks by selecting the `block` column from the `PermitEntity` repository,
+   * filtered by the provided `extractorId`
+   *
+   * @returns A query builder selecting used blocks
+   */
+  createUsedBlocksQuery = (): SelectQueryBuilder<PermitEntity> =>
+    this.actions.createUsedBlocksQuery(this.getId());
 }
 
 export default PermitExtractor;