+++ packages/scanners/evm-scanner/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rosen-bridge/evm-scanner
 
+## 1.0.0
+
+### Major Changes
+
+- Update AbstractExtractor interface; use `BlockInfo` in `processTransactions` and rename `initializeBoxes` to `initializeData`
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@1.0.0
+
 ## 0.1.4
 
 ### Patch Changes